requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.23
blinker==1.7.0
itsdangerous==2.1.2
Pillow